refactor: simplify providers
